requirements.txt
streamlit==1.32.0
requests==2.31.0
fpdf2==2.7.9
python-pptx==0.6.23
google-genai>=1.0.0